Get rid of some warnings

+++ src/graphics/sp/sp_base.cpp
@@ -451,7 +451,7 @@ void init()
         int skinning_tbo_limit;
         glGetIntegerv(GL_MAX_TEXTURE_BUFFER_SIZE_ARB, &skinning_tbo_limit);
         
-        g_skinning_use_tbo = skinning_tbo_limit >= STKConfig::get()->m_max_skinning_bones << 6;
+        g_skinning_use_tbo = skinning_tbo_limit >= (int)(STKConfig::get()->m_max_skinning_bones << 6);
     }
     else
     {
